requirements.txt
numpy~=1.24.2
scipy>=1.3.2
probreg~=0.3.6
scikit-learn~=1.2.2
open3d~=0.17.0
tifffile~=2023.3.15
plotly~=5.13.1
matplotlib~=3.7.1
pillow~=9.4.0
scikit-image~=0.20.0
numpy-stl~=3.1.1
pyamg~=5.0.1
